Make some imports relative
Move date range generator to separate module.
Simpler ordering for DetectorHisparc
