+++ django_publicdb/histograms/esd.py
@@ -12,7 +12,7 @@ from sapphire.analysis import process_events, coincidences, calibration
 from sapphire import clusters
 
 from ..inforecords.models import Station
-import datastore
+from . import datastore
 
 from django.conf import settings
