+++ django_publicdb/inforecords/models.py
@@ -308,7 +308,7 @@ class DetectorHisparc(models.Model):
     class Meta:
         unique_together = [('station', 'startdate')]
         verbose_name_plural = 'Detector HiSPARC'
-        ordering = ('station__number',)
+        ordering = ('station',)
 
 
 class ElectronicsType(models.Model):
